feat: add session metrics tracking and update related functions (#1201)
* feat: add session metrics tracking and update related functions

* feat: implement session metrics API with filtering and aggregation

* feat: add index for org_id in session_metrics table

+++ gateway/models/session.go
@@ -494,7 +494,7 @@ func UpdateSessionIntegrationMetadata(orgID, sid string, metadata map[string]any
 	return res.Error
 }
 
-func UpdateSessionAnalyzerMetrics(orgID, sid string, metrics map[string]int) error {
+func UpdateSessionAnalyzerMetrics(orgID, sid string, metrics map[string]int64) error {
 	res := DB.Table("private.sessions AS s").
 		Where("s.id = ? AND s.org_id = ?", sid, orgID).
 		Update("metrics", gorm.Expr(`
